+++ Quiz2/Controllers/QuestionController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Quiz2.Data;
 using Quiz2.DTO;
+using Quiz2.Helper;
 using Quiz2.Models;
 using Quiz2.Services;
 
@@ -15,10 +16,12 @@ namespace Quiz2.Controllers
     public class QuestionController: ControllerBase
     {
         private readonly IQuestionService questionService;
+        private readonly IQuizService quizService;
 
-        public QuestionController(IQuestionService questionService)
+        public QuestionController(IQuestionService questionService, IQuizService quizService)
         {
             this.questionService = questionService;
+            this.quizService = quizService;
         }
 
         // GET: api/Question/5
@@ -39,15 +42,29 @@ namespace Quiz2.Controllers
         [HttpPatch("{questionId}")]
         public ActionResult<Question> UpdateQuestion(int questionId, UpdateQuestionDto updateQuestionDto)
         {
-            return questionService.UpdateQuestion(questionId, updateQuestionDto);
+            if (UserIdCheck(questionId))
+            {
+                return questionService.UpdateQuestion(questionId, updateQuestionDto);
+            }
+            return null;
         }
 
         //DELETE: api/Question/5
         [HttpDelete("{questionId}")]
         public IActionResult DeleteQuestion(int questionId)
         {
-            questionService.DeleteQuestion(questionId);
+            if (UserIdCheck(questionId))
+            {
+                questionService.DeleteQuestion(questionId);   
+            }
             return NoContent();
+        }
+        
+        private bool UserIdCheck(int questionId)
+        {
+            var question = questionService.GetQuestion(questionId);
+            var quiz = quizService.GetQuiz(question.QuizId);
+            return quiz.Owner.Id == HttpContext.User.GetUserId();
         }
     }
 }
